server/requirements.txt
Flask==2.3.3
flask-cors==4.0.0

together==0.2.11
python-dotenv==1.0.1
google-generativeai==0.3.2

langchain==0.1.14
langchain-community==0.0.31
langchain-huggingface==0.0.3

pypdf==3.17.4
python-docx==1.1.0